Streamlit/requirements.txt
streamlit
folium
geopandas
branca.colormap
streamlit_folium